+++ HumansData/HumanTypes.h
@@ -19,11 +19,11 @@ enum class HumanType
 };
 struct HumanBase
 {
-    std::vector<uint32_t> posX;
-    std::vector<uint32_t> posY;
-    std::vector<uint32_t> targetX;
-    std::vector<uint32_t> targetY;
-    std::vector<uint16_t> points;
-    std::vector<uint16_t> moves;
+    std::vector<uint16_t> posX;
+    std::vector<uint16_t> posY;
+    std::vector<uint16_t> targetX;
+    std::vector<uint16_t> targetY;
+    std::vector<uint8_t> points;
+    std::vector<uint8_t> moves;
     std::vector<BuildingType> targetBuilding;
 };
