+++ XY/XY.h
@@ -2,6 +2,6 @@
 #include <cstdint>
 struct XY
 {
-    uint32_t x;
-    uint32_t y;
+    uint16_t x;
+    uint16_t y;
 };
